Restructuring the src/components folder
