+++ src/components/AmbientStereoSound.tsx
