+++ src/scenes/forest/Forest.tsx
@@ -2,8 +2,8 @@ import { Canvas } from '@react-three/fiber'
 import { Loader, PointerLockControls, Stars } from '@react-three/drei'
 import { Suspense } from 'react'
 import { Physics } from '@react-three/cannon'
-import { Player } from '../../components/organisms/Player'
-import { Ground } from '../../components/atoms/Ground'
+import { Player } from '../../components/Player'
+import { Ground } from '../../components/Ground'
 
 import Collider from '../../scenes/forest/components/Collider'
 import BaseRock from '../../scenes/forest/components/BaseRock'
