+++ src/scenes/gorge/Gorge.tsx
@@ -2,9 +2,9 @@ import { Canvas } from '@react-three/fiber'
 import { Loader, PointerLockControls, Stars } from '@react-three/drei'
 import { Suspense } from 'react'
 import { Physics } from '@react-three/cannon'
-import { Player } from '../../components/organisms/Player'
-import { Ground } from '../../components/atoms/Ground'
-import Water from '../../components/atoms/TaminaWater'
+import { Player } from '../../components/Player'
+import { Ground } from '../../components/Ground'
+import Water from './components/Water'
 
 import Collider from '../../scenes/gorge/components/Collider'
 import Lamps from '../../scenes/gorge/components/Lamps'
